fix usecase
